+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-  - New credential type "ssh-key" 
+  - New credential type "ssh-key"
+  - New action "add-ssh-key" for NuvlaBox resource
+  - New action "revoke-ssh-key" for NuvlaBox resource 
  
 ### Changed